finished chat feature

+++ HikingPal/app/src/main/java/com/cpen391/module2/hikingpal/module/Message.java
@@ -7,7 +7,7 @@ public class Message {
     private String message;
     private int sender;
 
-    public Message(int id, String message, int sender) {
+    public Message(long id, String message, int sender) {
         this.id = id;
         this.message = message;
         this.sender = sender;
